# Adopt the Arm SIMD-optimized p384 fields to point operations