--- a/arm/proofs/bignum_montmul_p521_neon.ml
+++ b/arm/proofs/bignum_montmul_p521_neon.ml
@@ -701,9 +701,9 @@ let bignum_montmul_p521_interm1_core_mc =
 let BIGNUM_MONTMUL_P521_INTERM1_CORE_EXEC =
   ARM_MK_EXEC_RULE bignum_montmul_p521_interm1_core_mc;;
 
-let equiv_input_states = new_definition
+let montmul_p521_eqin = new_definition
   `!s1 s1' x y z stackpointer.
-    (equiv_input_states:(armstate#armstate)->int64->int64->int64->int64->bool)
+    (montmul_p521_eqin:(armstate#armstate)->int64->int64->int64->int64->bool)
         (s1,s1') x y z stackpointer <=>
     (C_ARGUMENTS [z; x; y] s1 /\
      C_ARGUMENTS [z; x; y] s1' /\
@@ -714,9 +714,9 @@ let equiv_input_states = new_definition
      (?b. bignum_from_memory (y,9) s1 = b /\
           bignum_from_memory (y,9) s1' = b))`;;
 
-let equiv_output_states = new_definition
+let montmul_p521_eqout = new_definition
   `!s1 s1' z stackpointer.
-    (equiv_output_states:(armstate#armstate)->int64->int64->bool)
+    (montmul_p521_eqout:(armstate#armstate)->int64->int64->bool)
         (s1,s1') z stackpointer <=>
     (?a.
       read SP s1 = stackpointer /\
@@ -755,8 +755,8 @@ let equiv_goal1 = mk_equiv_statement_simple
        [(word pc,LENGTH bignum_montmul_p521_core_mc);
         (word pc2,LENGTH bignum_montmul_p521_interm1_core_mc);
         (z,8 * 9); (x:int64,8 * 9); (y:int64,8 * 9)]`
-    equiv_input_states
-    equiv_output_states
+    montmul_p521_eqin
+    montmul_p521_eqout
     bignum_montmul_p521_core_mc
     `MAYCHANGE [PC; X3; X4; X5; X6; X7; X8; X9;
                 X10; X11; X12; X13; X14; X15; X16; X17; X19;
@@ -765,8 +765,11 @@ let equiv_goal1 = mk_equiv_statement_simple
      MAYCHANGE [memory :> bignum(z,9);
                 memory :> bytes(stackpointer,80)]`
     bignum_montmul_p521_interm1_core_mc
-    `MAYCHANGE_REGS_AND_FLAGS_PERMITTED_BY_ABI ,,
-     MAYCHANGE [X19; X20; X21; X22; X23; X24; X25; X26] ,,
+    `MAYCHANGE [PC; X1; X2; X3; X4; X5; X6; X7; X8; X9;
+                X10; X11; X12; X13; X14; X15; X16; X17; X19;
+                X20; X21; X22; X23; X24; X25; X26] ,,
+     MAYCHANGE MODIFIABLE_SIMD_REGS ,,
+     MAYCHANGE SOME_FLAGS ,,
      MAYCHANGE [memory :> bignum(z,9);
                 memory :> bytes(stackpointer,80)]`;;
 
@@ -779,12 +782,12 @@ extra_word_CONV :=
 let BIGNUM_MONTMUL_P521_CORE_EQUIV1 = time prove(equiv_goal1,
 
   REWRITE_TAC[MAYCHANGE_REGS_AND_FLAGS_PERMITTED_BY_ABI;SOME_FLAGS;
-    ALLPAIRS;ALL;NONOVERLAPPING_CLAUSES;
+    MODIFIABLE_SIMD_REGS;ALLPAIRS;ALL;NONOVERLAPPING_CLAUSES;
     fst BIGNUM_MONTMUL_P521_CORE_EXEC;
     fst BIGNUM_MONTMUL_P521_INTERM1_CORE_EXEC] THEN
   REPEAT STRIP_TAC THEN
   (** Initialize **)
-  EQUIV_INITIATE_TAC equiv_input_states THEN
+  EQUIV_INITIATE_TAC montmul_p521_eqin THEN
   REPEAT (FIRST_X_ASSUM BIGNUM_EXPAND_AND_DIGITIZE_TAC) THEN
   ASM_PROPAGATE_DIGIT_EQS_FROM_EXPANDED_BIGNUM_TAC THEN
   (* necessary to run ldr qs *)
@@ -800,7 +803,7 @@ let BIGNUM_MONTMUL_P521_CORE_EQUIV1 = time prove(equiv_goal1,
   ASM_REWRITE_TAC[] THEN
   REPEAT CONJ_TAC THENL [
     (** SUBGOAL 1. Outputs **)
-    ASM_REWRITE_TAC[equiv_output_states;mk_equiv_regs;mk_equiv_bool_regs;
+    ASM_REWRITE_TAC[montmul_p521_eqout;mk_equiv_regs;mk_equiv_bool_regs;
                     BIGNUM_EXPAND_CONV `bignum_from_memory (ptr,9) state`;
                     C_ARGUMENTS] THEN
     REPEAT (HINT_EXISTS_REFL_TAC THEN ASM_REWRITE_TAC[]);
@@ -848,16 +851,22 @@ let equiv_goal2 = mk_equiv_statement_simple
        [(word pc,LENGTH bignum_montmul_p521_interm1_core_mc);
         (word pc2,LENGTH bignum_montmul_p521_neon_core_mc);
         (z,8 * 9); (x:int64,8 * 9); (y:int64,8 * 9)]`
-    equiv_input_states
-    equiv_output_states
+    montmul_p521_eqin
+    montmul_p521_eqout
     bignum_montmul_p521_interm1_core_mc
-    `MAYCHANGE_REGS_AND_FLAGS_PERMITTED_BY_ABI ,,
-     MAYCHANGE [X19; X20; X21; X22; X23; X24; X25; X26] ,,
+    `MAYCHANGE [PC; X1; X2; X3; X4; X5; X6; X7; X8; X9;
+                X10; X11; X12; X13; X14; X15; X16; X17; X19;
+                X20; X21; X22; X23; X24; X25; X26] ,,
+     MAYCHANGE MODIFIABLE_SIMD_REGS ,,
+     MAYCHANGE SOME_FLAGS ,,
      MAYCHANGE [memory :> bignum(z,9);
                 memory :> bytes(stackpointer,80)]`
     bignum_montmul_p521_neon_core_mc
-    `MAYCHANGE_REGS_AND_FLAGS_PERMITTED_BY_ABI ,,
-     MAYCHANGE [X19; X20; X21; X22; X23; X24; X25; X26] ,,
+    `MAYCHANGE [PC; X1; X2; X3; X4; X5; X6; X7; X8; X9;
+                X10; X11; X12; X13; X14; X15; X16; X17; X19;
+                X20; X21; X22; X23; X24; X25; X26] ,,
+     MAYCHANGE MODIFIABLE_SIMD_REGS ,,
+     MAYCHANGE SOME_FLAGS ,,
      MAYCHANGE [memory :> bignum(z,9);
                 memory :> bytes(stackpointer,80)]`;;
 
@@ -874,12 +883,12 @@ let BIGNUM_MONTMUL_P521_CORE_EQUIV2 = time prove(
   equiv_goal2,
 
   REWRITE_TAC[MAYCHANGE_REGS_AND_FLAGS_PERMITTED_BY_ABI;SOME_FLAGS;
-    ALLPAIRS;ALL;NONOVERLAPPING_CLAUSES;
+    MODIFIABLE_SIMD_REGS;ALLPAIRS;ALL;NONOVERLAPPING_CLAUSES;
     fst BIGNUM_MONTMUL_P521_INTERM1_CORE_EXEC;
     fst BIGNUM_MONTMUL_P521_NEON_CORE_EXEC] THEN
   REPEAT STRIP_TAC THEN
   (** Initialize **)
-  EQUIV_INITIATE_TAC equiv_input_states THEN
+  EQUIV_INITIATE_TAC montmul_p521_eqin THEN
   REPEAT (FIRST_X_ASSUM BIGNUM_EXPAND_AND_DIGITIZE_TAC) THEN
   ASM_PROPAGATE_DIGIT_EQS_FROM_EXPANDED_BIGNUM_TAC THEN
   (* necessary to run ldr qs *)
@@ -898,7 +907,7 @@ let BIGNUM_MONTMUL_P521_CORE_EQUIV2 = time prove(
   ASM_REWRITE_TAC[] THEN
   REPEAT CONJ_TAC THENL [
     (** SUBGOAL 1. Outputs **)
-    ASM_REWRITE_TAC[equiv_output_states;mk_equiv_regs;mk_equiv_bool_regs;
+    ASM_REWRITE_TAC[montmul_p521_eqout;mk_equiv_regs;mk_equiv_bool_regs;
                     BIGNUM_EXPAND_CONV `bignum_from_memory (ptr,9) state`;
                     C_ARGUMENTS] THEN
     REPEAT (HINT_EXISTS_REFL_TAC THEN ASM_REWRITE_TAC[]);
@@ -927,8 +936,8 @@ let equiv_goal = mk_equiv_statement_simple
        [(word pc,LENGTH bignum_montmul_p521_core_mc);
         (word pc2,LENGTH bignum_montmul_p521_neon_core_mc);
         (z,8 * 9); (x:int64,8 * 9); (y:int64,8 * 9)]`
-    equiv_input_states
-    equiv_output_states
+    montmul_p521_eqin
+    montmul_p521_eqout
     bignum_montmul_p521_core_mc
     `MAYCHANGE [PC; X3; X4; X5; X6; X7; X8; X9;
                 X10; X11; X12; X13; X14; X15; X16; X17; X19;
@@ -937,17 +946,20 @@ let equiv_goal = mk_equiv_statement_simple
      MAYCHANGE [memory :> bignum(z,9);
                 memory :> bytes(stackpointer,80)]`
     bignum_montmul_p521_neon_core_mc
-    `MAYCHANGE_REGS_AND_FLAGS_PERMITTED_BY_ABI ,,
-     MAYCHANGE [X19; X20; X21; X22; X23; X24; X25; X26] ,,
+    `MAYCHANGE [PC; X1; X2; X3; X4; X5; X6; X7; X8; X9;
+                X10; X11; X12; X13; X14; X15; X16; X17; X19;
+                X20; X21; X22; X23; X24; X25; X26] ,,
+     MAYCHANGE MODIFIABLE_SIMD_REGS ,,
+     MAYCHANGE SOME_FLAGS ,,
      MAYCHANGE [memory :> bignum(z,9);
                 memory :> bytes(stackpointer,80)]`;;
 
-let equiv_output_states_TRANS = prove(
+let montmul_p521_eqout_TRANS = prove(
   `!s s2 s' z stackpointer.
-    equiv_output_states (s,s') z stackpointer/\
-    equiv_output_states (s',s2) z stackpointer
-    ==> equiv_output_states (s,s2) z stackpointer`,
-  MESON_TAC[equiv_output_states]);;
+    montmul_p521_eqout (s,s') z stackpointer/\
+    montmul_p521_eqout (s',s2) z stackpointer
+    ==> montmul_p521_eqout (s,s2) z stackpointer`,
+  MESON_TAC[montmul_p521_eqout]);;
 
 let BIGNUM_MONTMUL_P521_CORE_EQUIV = time prove(equiv_goal,
   REPEAT STRIP_TAC THEN
@@ -980,49 +992,11 @@ let BIGNUM_MONTMUL_P521_CORE_EQUIV = time prove(equiv_goal,
   ] THEN
   STRIP_TAC THEN
 
-  (* instantiate first equiv *)
-  ENSURES2_TRANS_TAC BIGNUM_MONTMUL_P521_CORE_EQUIV1 BIGNUM_MONTMUL_P521_CORE_EQUIV2 THEN
-
-  (* break 'ALL nonoverlapping' in assumptions *)
-  RULE_ASSUM_TAC (REWRITE_RULE[
-      ALLPAIRS;ALL;
-      fst BIGNUM_MONTMUL_P521_CORE_EXEC;
-      fst BIGNUM_MONTMUL_P521_NEON_CORE_EXEC;
-      fst BIGNUM_MONTMUL_P521_INTERM1_CORE_EXEC;
-      NONOVERLAPPING_CLAUSES]) THEN
-  REPEAT SPLIT_FIRST_CONJ_ASSUM_TAC THEN
-
-  MATCH_MP_TAC ENSURES2_WEAKEN THEN
-  REWRITE_TAC[] THEN
-  REPEAT CONJ_TAC THENL [
-    REPEAT STRIP_TAC THEN ASM_REWRITE_TAC[] THEN
-    REWRITE_TAC[TAUT `(p /\ q /\ r) /\ p /\ q /\ r' <=> p /\ q /\ r /\ r'`] THEN
-    EXISTS_TAC
-      `write (memory :> bytelist
-          (word pc3,LENGTH bignum_montmul_p521_interm1_core_mc))
-          bignum_montmul_p521_interm1_core_mc
-          (write PC (word pc3) s')` THEN
-    PROVE_CONJ_OF_EQ_READS_TAC BIGNUM_MONTMUL_P521_INTERM1_CORE_EXEC THENL [
-      UNDISCH_TAC `equiv_input_states (s,s') x y z stackpointer` THEN
-      REWRITE_TAC[equiv_input_states;C_ARGUMENTS;BIGNUM_FROM_MEMORY_BYTES;
-          fst BIGNUM_MONTMUL_P521_INTERM1_CORE_EXEC] THEN
-      STRIP_TAC THEN ASM_REWRITE_TAC[] THEN
-      REPEAT (TRY HINT_EXISTS_REFL_TAC THEN
-        PROVE_CONJ_OF_EQ_READS_TAC BIGNUM_MONTMUL_P521_INTERM1_CORE_EXEC);
-
-      UNDISCH_TAC `equiv_input_states (s,s') x y z stackpointer` THEN
-      REWRITE_TAC[equiv_input_states;C_ARGUMENTS;BIGNUM_FROM_MEMORY_BYTES;
-          fst BIGNUM_MONTMUL_P521_INTERM1_CORE_EXEC] THEN
-      STRIP_TAC THEN ASM_REWRITE_TAC[] THEN
-      REPEAT (TRY HINT_EXISTS_REFL_TAC THEN
-        PROVE_CONJ_OF_EQ_READS_TAC BIGNUM_MONTMUL_P521_INTERM1_CORE_EXEC);
-    ];
-
-    REPEAT GEN_TAC THEN STRIP_TAC THEN
-    ASM_REWRITE_TAC[] THEN ASM_MESON_TAC[equiv_output_states_TRANS];
-
-    SUBSUMED_MAYCHANGE_TAC
-  ]);;
+  EQUIV_TRANS_TAC
+    (BIGNUM_MONTMUL_P521_CORE_EQUIV1,BIGNUM_MONTMUL_P521_CORE_EQUIV2)
+    (montmul_p521_eqin,montmul_p521_eqout_TRANS)
+    (BIGNUM_MONTMUL_P521_CORE_EXEC,BIGNUM_MONTMUL_P521_INTERM1_CORE_EXEC,
+     BIGNUM_MONTMUL_P521_NEON_CORE_EXEC));;
 
 
 
@@ -1092,10 +1066,13 @@ let BIGNUM_MONTMUL_P521_NEON_CORE_CORRECT = prove
                   (a < p_521 /\ b < p_521
                    ==> bignum_from_memory (z,9) s =
                         (inverse_mod p_521 (2 EXP 576) * a * b) MOD p_521))
-             (MAYCHANGE_REGS_AND_FLAGS_PERMITTED_BY_ABI ,,
-              MAYCHANGE [X19; X20; X21; X22; X23; X24; X25; X26] ,,
+             (MAYCHANGE [PC; X1; X2; X3; X4; X5; X6; X7; X8; X9;
+                         X10; X11; X12; X13; X14; X15; X16; X17; X19;
+                         X20; X21; X22; X23; X24; X25; X26] ,,
+              MAYCHANGE MODIFIABLE_SIMD_REGS ,,
+              MAYCHANGE SOME_FLAGS ,,
               MAYCHANGE [memory :> bignum(z,9);
-                         memory :> bytes(stackpointer,80)])`,
+                          memory :> bytes(stackpointer,80)])`,
   REPEAT GEN_TAC THEN
   (* Prepare pc for the original program.  *)
   SUBGOAL_THEN
@@ -1115,43 +1092,10 @@ let BIGNUM_MONTMUL_P521_NEON_CORE_CORRECT = prove
 
   REPEAT_N 2 STRIP_TAC THEN
 
-  VCGEN_EQUIV_TAC BIGNUM_MONTMUL_P521_CORE_EQUIV BIGNUM_MONTMUL_P521_CORE_CORRECT_N
-    [fst BIGNUM_MONTMUL_P521_CORE_EXEC;fst BIGNUM_MONTMUL_P521_NEON_CORE_EXEC] THEN
-
-  (* unfold definitions that may block tactics *)
-  RULE_ASSUM_TAC (REWRITE_RULE[ALL;NONOVERLAPPING_CLAUSES;
-      fst BIGNUM_MONTMUL_P521_EXEC;
-      fst BIGNUM_MONTMUL_P521_NEON_EXEC]) THEN
-  REPEAT SPLIT_FIRST_CONJ_ASSUM_TAC THEN
-  REWRITE_TAC[C_ARGUMENTS;BIGNUM_FROM_MEMORY_BYTES] THEN
-  REPEAT CONJ_TAC THENL [
-    (** SUBGOAL 1. Precond **)
-    X_GEN_TAC `s2:armstate` THEN REPEAT STRIP_TAC THEN
-    SUBGOAL_THEN `4 divides val (word pc2:int64)` ASSUME_TAC THENL
-    [ FIRST_ASSUM (fun th ->
-        MP_TAC th THEN REWRITE_TAC[DIVIDES_4_VAL_WORD_64;aligned_bytes_loaded_word]
-        THEN METIS_TAC[]) THEN NO_TAC; ALL_TAC ] THEN
-    ASM_REWRITE_TAC[equiv_input_states;C_ARGUMENTS] THEN
-    EXISTS_TAC
-      `write (memory :> bytelist
-          (word pc,LENGTH (APPEND bignum_montmul_p521_core_mc barrier_inst_bytes)))
-          (APPEND bignum_montmul_p521_core_mc barrier_inst_bytes)
-          (write PC (word pc) s2)` THEN
-    (* Expand variables appearing in the equiv relation *)
-    REPEAT CONJ_TAC THEN
-    TRY (PROVE_CONJ_OF_EQ_READS_TAC BIGNUM_MONTMUL_P521_CORE_EXEC) THEN
-    (* Now has only one subgoal: the '?a. ...' part of input equivalence! *)
-    REWRITE_TAC[BIGNUM_FROM_MEMORY_BYTES] THEN
-    REPEAT (HINT_EXISTS_REFL_TAC THEN
-      PROVE_CONJ_OF_EQ_READS_TAC BIGNUM_MONTMUL_P521_CORE_EXEC);
-
-    (** SUBGOAL 2. Postcond **)
-    MESON_TAC[equiv_output_states;BIGNUM_FROM_MEMORY_BYTES;
-        fst BIGNUM_MONTMUL_P521_NEON_CORE_EXEC];
-
-    (** SUBGOAL 3. Frame **)
-    MESON_TAC[MAYCHANGE_REGS_AND_FLAGS_PERMITTED_BY_ABI]
-  ]);;
+  PROVE_ENSURES_FROM_EQUIV_AND_ENSURES_N_TAC
+    BIGNUM_MONTMUL_P521_CORE_EQUIV BIGNUM_MONTMUL_P521_CORE_CORRECT_N
+    (BIGNUM_MONTMUL_P521_CORE_EXEC,BIGNUM_MONTMUL_P521_NEON_CORE_EXEC)
+    (montmul_p521_eqin,montmul_p521_eqout));;
 
 
 let BIGNUM_MONTMUL_P521_NEON_CORRECT = prove
@@ -1172,10 +1116,13 @@ let BIGNUM_MONTMUL_P521_NEON_CORRECT = prove
                   (a < p_521 /\ b < p_521
                    ==> bignum_from_memory (z,9) s =
                         (inverse_mod p_521 (2 EXP 576) * a * b) MOD p_521))
-             (MAYCHANGE_REGS_AND_FLAGS_PERMITTED_BY_ABI ,,
-              MAYCHANGE [X19; X20; X21; X22; X23; X24; X25; X26] ,,
+             (MAYCHANGE [PC; X1; X2; X3; X4; X5; X6; X7; X8; X9;
+                         X10; X11; X12; X13; X14; X15; X16; X17; X19;
+                         X20; X21; X22; X23; X24; X25; X26] ,,
+              MAYCHANGE MODIFIABLE_SIMD_REGS ,,
+              MAYCHANGE SOME_FLAGS ,,
               MAYCHANGE [memory :> bignum(z,9);
-                         memory :> bytes(stackpointer,80)])`,
+                          memory :> bytes(stackpointer,80)])`,
 
   ARM_SUB_LIST_OF_MC_TAC BIGNUM_MONTMUL_P521_NEON_CORE_CORRECT
       bignum_montmul_p521_neon_core_mc_def